# services/stash-service/api-spec/src/main/resources/openapi/stash-api.yaml
openapi: 3.0.3
info:
  title: Linkpouch Stash Service API
  description: API for managing stashes and links in Linkpouch
  version: 1.0.0
  contact:
    name: Linkpouch Team
servers:
  - url: /api
    description: Local development
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT issued by POST /stashes/{stashId}/access-token
  schemas:
...
    FolderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stashId:
          type: string
          format: uuid
        parentFolderId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateFolderRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        parentFolderId:
          type: string
          format: uuid
          nullable: true
    RenameFolderRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
    MoveFolderRequest:
      type: object
      properties:
        newParentFolderId:
          type: string
          format: uuid
          nullable: true
        insertAfterId:
          type: string
          format: uuid
          nullable: true
    MoveLinkToFolderRequest:
      type: object
      properties:
        folderId:
          type: string
          format: uuid
          nullable: true
          description: Target folder ID. Null moves the link to the root.
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        errorCode:
          type: string
          description: Machine-readable error code (e.g. PASSWORD_REQUIRED, INVALID_TOKEN)
        path:
          type: string
        signatureRefreshedAt:
          type: string
          format: date-time
          description: When the stash signature was last regenerated (only set when errorCode is SIGNATURE_REGENERATED)
...
paths:
...
  /stashes/{stashId}/folders:
    parameters:
      - name: stashId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List all folders in a stash
      operationId: listFolders
      tags:
        - Folders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FolderResponse'
        '401':
          description: Invalid or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a folder in a stash
      operationId: createFolder
      tags:
        - Folders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        '201':
          description: Folder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stashes/{stashId}/folders/{folderId}:
    parameters:
      - name: stashId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: folderId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Rename a folder
      operationId: renameFolder
      tags:
        - Folders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameFolderRequest'
      responses:
        '200':
          description: Folder renamed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderResponse'
        '401':
          description: Invalid or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Folder does not belong to this stash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Folder not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a folder (and recursively all sub-folders; links moved to root)
      operationId: deleteFolder
      tags:
        - Folders
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Folder deleted
        '401':
          description: Invalid or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Folder does not belong to this stash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Folder not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stashes/{stashId}/folders/{folderId}/move:
    parameters:
      - name: stashId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: folderId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      summary: Move a folder to a new parent (and optionally reorder)
      operationId: moveFolder
      tags:
        - Folders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveFolderRequest'
      responses:
        '204':
          description: Folder moved
        '401':
          description: Invalid or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Folder does not belong to this stash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Folder not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stashes/{stashId}/links/{linkId}/folder:
    parameters:
      - name: stashId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: linkId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Move a link to a different folder (or root)
      operationId: moveLinkToFolder
      tags:
        - Folders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveLinkToFolderRequest'
      responses:
        '204':
          description: Link moved
        '401':
          description: Invalid or expired access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Link does not belong to this stash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Link not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...